cc: allow git responses
This commit allows users to specify a Git repo for a cc group. Specifying a repo with an optional subdirectory clones the repo locally and begins a 30 second rotating git pull command for the local repository. If a subdirectory is specified, automatic git pull is unsupported. Instead use the git pull button on any custom command to repeat the process which occurs when a repo is first defined for a cc group. When a repository (and optional subdirectory) is specified, ccs within the group gain the ability to get their responses from a file within that repo. That file must follow the structure <guildID>/<cmdID>.yag. You may toggle any command to get its response from the repository specified, and click git pull on a cc's edit page to refresh the repository.
Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>